+++ internal/resp/write.go
@@ -84,7 +84,7 @@ func WriteRespType(conn io.Writer, val RespType) error {
 	}
 }
 
-func WriteInt(conn io.Writer, i int) error {
+func WriteInt(conn io.Writer, i int64) error {
 	str := fmt.Sprintf(":%d\r\n", i)
 	_, err := conn.Write([]byte(str))
 	return err
